Conflicts:
	contrib/quix/lib/quix/diagnostic.rb
	contrib/quix/lib/quix/kernel.rb
	contrib/quix/lib/quix/thread_local.rb
	contrib/quix/lib/quix/vars.rb

+++ contrib/quix/lib/quix/fileutils.rb
@@ -1,37 +1,33 @@
 
 require 'tmpdir'
-require 'quix/builtin/kernel/tap'
+require 'quix/kernel'
 
-module Quix
-  module FileUtils
-    def rename_file(file, new_name)
-      #
-      # For case-insensitive systems, we must move the file elsewhere
-      # before changing case.
-      #
-      temp = File.join(Dir.tmpdir, File.basename(file))
-      ::FileUtils.mv(file, temp)
-      begin
-        ::FileUtils.mv(temp, new_name)
-      rescue
-        ::FileUtils.mv(temp, file)
-        raise
-      end
+module FileUtils
+  module_function
+
+  def rename_file(file, new_name)
+    #
+    # For case-insensitive systems, we must move the file elsewhere
+    # before changing case.
+    #
+    temp = File.join(Dir.tmpdir, File.basename(file))
+    mv(file, temp)
+    begin
+      mv(temp, new_name)
+    rescue
+      mv(temp, file)
+      raise
     end
-    
-    def replace_file(file)
-      old_contents = File.read(file)
-      yield(old_contents).tap { |new_contents|
+  end
+  
+  def replace_file(file, opts = {})
+    old_contents = File.read(file)
+    yield(old_contents).tap { |new_contents|
+      if opts[:force] or old_contents != new_contents
         File.open(file, "w") { |output|
           output.print(new_contents)
         }
-      }
-    end
-    
-    def stem(file)
-      file.sub(%r!#{File.extname(file)}\Z!, "")
-    end
-
-    extend self
+      end
+    }
   end
 end
